aioc-hospital-login-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.2.1
PyJWT==2.8.0
bcrypt==4.2.1
python-multipart==0.0.9
alembic==1.13.1
